load the table view of friends #742

+++ Project/LooseLeaf/MMCloudKitBaseState.m
@@ -44,7 +44,7 @@
         [[SPRSimpleCloudKitManager sharedManager] silentlyVerifyiCloudAccountStatusOnComplete:^(SCKMAccountStatus accountStatus,
                                                                                                 SCKMApplicationPermissionStatus permissionStatus,
                                                                                                 NSError *error) {
-            NSLog(@"got account status and permisison info %d %d %p!", accountStatus, permissionStatus, error);
+            NSLog(@"got account status and permisison info %d %d %p!", (int) accountStatus, (int) permissionStatus, error);
             @synchronized(self){
                 isCheckingStatus = NO;
             }
